streamlit/llama-index/requirements.txt
streamlit==1.22.0
llama-index==0.5.27
python-dotenv
